alert on delete,navigation goback, style optimization

+++ src/screens/user/UserProductsScreen.js
@@ -1,5 +1,5 @@
 import React from "react";
-import { View, Text, FlatList, Button } from "react-native";
+import { View, Text, FlatList, Button, Alert } from "react-native";
 import { useSelector, useDispatch } from "react-redux";
 import { HeaderButtons, Item } from "react-navigation-header-buttons";
 
@@ -13,7 +13,16 @@ const UserProductsScreen = ({ navigation }) => {
   const dispatch = useDispatch();
 
   const deleteHandler = productId => {
-    dispatch(actions.deleteProduct(productId));
+    Alert.alert("Delete item", "Are you really want to delete this item?", [
+      { text: "No", style: "default" },
+      {
+        text: "Yes",
+        style: "destructive",
+        onPress: () => {
+          dispatch(actions.deleteProduct(productId));
+        }
+      }
+    ]);
   };
 
   const editHandler = productId => {
